modbus: move MODBUS over serial line code to own file
Move MODBUS over serial line code to own source file.
